Add Encryption class
