+++ RubSort/ApiApplication/Views/Authentication/ResetPassword.cshtml
@@ -1,19 +1,50 @@
-﻿@page
-@model ApiApplication.Views.Authentication.ResetPassword
-
-@{
-    Layout = null;
-}
+﻿@model ApiApplication.Models.RegisterModel
 
 <!DOCTYPE html>
 
 <html>
 <head>
-    <title></title>
+    <title>Сброс пароля</title>
 </head>
 <body>
-<div>
-    
-</div>
+    <div class="form-container">
+        <form asp-action="ResetPassword" asp-controller="Authentication" asp-anti-forgery="true">
+            <div class="validation" asp-validation-summary="ModelOnly" />
+            @Html.ValidationSummary(true, "", new { @class = "text-danger" })
+            <div>
+                @Html.ValidationMessageFor(model => model.Email)
+                <div class="form-group">
+                    <div class="input-group mb-3">
+                        <span asp-for="Email" id="inputGroup-sizing-default">Email</span>
+                        <input type="text"  asp-for="Email" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default">      
+                    </div>
+                    <span asp-validation-for="Email" />    
+                </div>
+
+                @Html.ValidationSummary(true, "", new { @class = "text-danger" })
+                <div>
+                    <div class="form-group">
+                        <div class="input-group mb-3">
+                            <span asp-for="Password" class="input-group-text" id="inputGroup-sizing-default">Password</span>
+                            <input type="password"  asp-for="Password" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default">      
+                        </div>
+                        <span asp-validation-for="Password" />    
+                    </div>
+            
+                    <div class="form-group">
+                        <div class="input-group mb-3">
+                            <span asp-for="ConfirmPassword" class="input-group-text" id="inputGroup-sizing-default">ConfirmPassword</span>
+                            <input type="password"  asp-for="ConfirmPassword" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-default">      
+                        </div>
+                        <span asp-validation-for="ConfirmPassword"/>   
+                    </div>
+        
+                </div>
+                <div class="form-group buttons-block">
+                    <button type="submit" class="btn btn-primary">Сбросить пароль</button>
+                </div>
+            </div>
+        </form>
+    </div>
 </body>
 </html>
