+++ RubSort/IdentitySystem/AuthenticationManager.cs
@@ -1,14 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
-using System.Security.Cryptography;
-using System.Text;
 using DataStorageSystem;
 
 namespace IdentitySystem
 {
-    public class AuthenticationManager
+    public class AuthenticationManager : IAuthenticationManager<ClaimsIdentity>
     {
         private readonly IEntityRepository<UserDbo> _userRepository;
         public AuthenticationManager(IEntityRepository<UserDbo> userRepository)
@@ -23,14 +20,22 @@ namespace IdentitySystem
 
         public void ChangePassword(string userEmail, string password)
         {
-            _userRepository.Update(new UserDbo() { Email = userEmail, Password = EncryptPassword(password, userEmail) });
+            _userRepository.Update(new UserDbo()
+            {
+                Email = userEmail, 
+                Password = Encryption.EncryptPasswordMD5(password, userEmail)
+            });
         }
 
         public ClaimsIdentity Register(string userEmail, string password)
         {
             if(IsRegisteredUser(userEmail))
                 return null;
-            _userRepository.Add(new UserDbo() {Email = userEmail, Password = EncryptPassword(password, userEmail)});
+            _userRepository.Add(new UserDbo()
+            {
+                Email = userEmail, 
+                Password = Encryption.EncryptPasswordMD5(password, userEmail)
+            });
             return Login(userEmail);
         }
 
@@ -44,20 +49,6 @@ namespace IdentitySystem
             };
             return new ClaimsIdentity(claims, "ApplicationCookie", ClaimsIdentity.DefaultNameClaimType,
                 ClaimsIdentity.DefaultRoleClaimType);
-        }
-        
-        private string EncryptPassword(string password, string userEmail)
-        {
-            var pass = Encoding.UTF8.GetBytes(password);
-            var entropy = GetEntropy(password + userEmail);
-            var protectedPass = ProtectedData.Protect(pass, entropy, DataProtectionScope.LocalMachine);
-            return Convert.ToBase64String(protectedPass);
-        }
-        
-        private byte[] GetEntropy(string entropyString)
-        {
-            var md5 = MD5.Create();
-            return md5.ComputeHash(Encoding.UTF8.GetBytes(entropyString));
         }
     }
 }
