+++ RubSort/IdentitySystem/IAuthenticationManager.cs
@@ -1,0 +1,13 @@
+﻿namespace IdentitySystem
+{
+    public interface IAuthenticationManager<T>
+    {
+        bool IsRegisteredUser(string userEmail);
+
+        void ChangePassword(string userEmail, string password);
+
+        T Register(string userEmail, string password);
+
+        T Login(string userEmail);
+    }
+}
